also fix doctests

+++ lib/explorer/data_frame.ex
@@ -2747,7 +2747,7 @@ defmodule Explorer.DataFrame do
       #Explorer.DataFrame<
         Polars[3 x 2]
         a string ["a", "b", "c"]
-        b s64 [1, 4, 9]
+        b f64 [1.0, 4.0, 9.0]
       >
 
   It's possible to "reuse" a variable for different computations:
